Increase CLS_BINARY_IMAGE_RUNTIME_NODE_COUNT
Testing shows that the number of images loaded into the average app
process has grown since this code was written. Overshooting this limit
is expensive for startup performance because Firebase will log a warning
to disk each time it occurs. Providing a bit of headroom here should
avoid this happening for all but the largest apps.